# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoRigo AI Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses for the current user
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessRequest"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /businesses/{id}/insights:
    get:
      operationId: getBusinessInsights
      tags: [businesses]
      summary: Get sector-specific automation insights
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessInsights"

  # ───── API Keys ─────
...
  /agents/approvals/pending:
    get:
      operationId: listPendingApprovals
      tags: [agents]
      summary: List pending agent approval requests
      parameters:
        - name: businessId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentApproval"
  /agents/approvals/{actionId}/approve:
    post:
      operationId: approveAgentAction
      tags: [agents]
      summary: Approve and execute a pending agent action
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /agents/approvals/{actionId}/reject:
    post:
      operationId: rejectAgentAction
      tags: [agents]
      summary: Reject a pending agent action
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents for a business
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a custom agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}:
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}/run:
    post:
      operationId: runAgent
      tags: [agents]
      summary: Run an AI agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessId:
                  type: string
              required: [businessId]
      responses:
        "200":
          description: Agent run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResult"
  /agents/{agentId}/test-run:
    post:
      operationId: testRunAgent
      tags: [agents]
      summary: Test run agent in preview mode
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessId:
                  type: string
              required: [businessId]
      responses:
        "200":
          description: Test run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResult"
  /agents/{agentId}/logs:
    get:
      operationId: getAgentLogs
      tags: [agents]
      summary: Get logs for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Agent logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentLog"

  # ───── Tools ─────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]

    # ── Business ──
    Business:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sector:
          type: ["string", "null"]
        country:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, country, isActive, createdAt]
    CreateBusinessRequest:
      type: object
      properties:
        name:
          type: string
        sector:
          type: ["string", "null"]
        country:
          type: string
        isActive:
          type: ["boolean", "null"]
      required: [name, country]
    UpdateBusinessRequest:
      type: object
      properties:
        name:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    BusinessInsights:
      type: object
      properties:
        sector:
          type: string
        recommendations:
          type: array
          items:
            type: string
        automationOpportunities:
          type: array
          items:
            type: string
      required: [sector, recommendations, automationOpportunities]

    # ── API Keys ──
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        isActive:
          type: boolean
        isBuiltIn:
          type: boolean
        businessId:
          type: string
        scheduleType:
          type: string
          enum: [none, hourly, daily, weekly, custom]
        scheduleTime:
          type: ["string", "null"]
        scheduleDay:
          type: ["integer", "null"]
        scheduleInterval:
          type: ["integer", "null"]
        nextRunAt:
          type: ["string", "null"]
        lastRunAt:
          type: ["string", "null"]
        toolAccess:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, type, isActive, isBuiltIn, businessId, scheduleType, toolAccess, createdAt]
    CreateAgentRequest:
      type: object
      properties:
        businessId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        systemPrompt:
          type: string
        toolAccess:
          type: array
          items:
            type: string
      required: [businessId, name, systemPrompt, toolAccess]
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        systemPrompt:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        scheduleType:
          type: ["string", "null"]
          enum: [none, hourly, daily, weekly, custom, "null"]
        scheduleTime:
          type: ["string", "null"]
        scheduleDay:
          type: ["integer", "null"]
        scheduleInterval:
          type: ["integer", "null"]
        toolAccess:
          type: ["array", "null"]
          items:
            type: string
    AgentLog:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        summary:
          type: string
        createdAt:
          type: string
      required: [id, agentId, summary, createdAt]
    AgentRunResult:
      type: object
      properties:
        summary:
          type: string
        actions:
          type: array
          items:
            type: string
        isPreview:
          type: boolean
      required: [summary, actions, isPreview]
    AgentApproval:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        actionType:
          type: string
        actionDescription:
          type: string
        toolName:
          type: string
        functionName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        expiresAt:
          type: string
        createdAt:
          type: string
      required: [id, agentId, actionType, actionDescription, toolName, functionName, status, expiresAt, createdAt]

    # ── Tools ──
...